se modifica archivo rutaVentas, metodo crearVenta
